Extract classes into their own files
