+++ src/main/java/com/google/idea/perf/CallTree.kt
@@ -57,44 +57,3 @@ class MutableCallTree(
         children.values.forEach(MutableCallTree::clear)
     }
 }
-
-/** Encapsulates aggregate statistic for a single tracepoint. */
-class TracepointStats(
-    val tracepoint: Tracepoint,
-    var callCount: Long = 0,
-    var wallTime: Long = 0,
-    var maxWallTime: Long = 0
-)
-
-object TreeAlgorithms {
-    /**
-     * Computes aggregate statistics for each tracepoint,
-     * being careful not to double-count the time spent in recursive calls.
-     */
-    fun computeFlatTracepointStats(root: CallTree): List<TracepointStats> {
-        val allStats = mutableMapOf<Tracepoint, TracepointStats>()
-        val ancestors = mutableSetOf<Tracepoint>()
-
-        fun bfs(node: CallTree) {
-            val nonRecursive = node.tracepoint !in ancestors
-            val stats = allStats.getOrPut(node.tracepoint) { TracepointStats(node.tracepoint) }
-            stats.callCount += node.callCount
-            if (nonRecursive) {
-                stats.wallTime += node.wallTime
-                stats.maxWallTime = stats.maxWallTime.coerceAtLeast(node.maxWallTime)
-                ancestors.add(node.tracepoint)
-            }
-            for (child in node.children.values) {
-                bfs(child)
-            }
-            if (nonRecursive) {
-                ancestors.remove(node.tracepoint)
-            }
-        }
-
-        bfs(root)
-        assert(ancestors.isEmpty())
-
-        return allStats.values.toList()
-    }
-}
